requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.21
locust==2.23.0
numpy==1.26.4
psycopg2-binary==2.9.9
